dam-mcp/cloud_functions/drive_sync/requirements.txt
google-cloud-storage>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
functions-framework>=3.0.0
